requirements.txt
# SkyPredict - Flight Delay Prediction
# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning (optional)
# tensorflow>=2.15.0

# Explainability
shap>=0.43.0
lime>=0.2.0

# Data Processing
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Feature Engineering
feature-engine>=1.6.0
category-encoders>=2.6.0

# Imbalanced Data
imbalanced-learn>=0.11.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# ML Tracking
mlflow>=2.9.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.27.0

# Weather API
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
python-dateutil>=2.8.0
